requirements.txt
amqp==5.0.1
asgiref==3.3.0
backcall==0.2.0
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==3.1.2
django-cors-headers==3.5.0
django-environ==0.4.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-webpack-loader==0.7.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
drf-irelation==0.0.5
idna==2.10
install==1.3.4
ipython==7.18.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
kombu==5.0.2
loguru==0.5.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
parso==0.8.0
pickleshare==0.7.5
Pillow==8.0.0
prompt-toolkit==3.0.8
Pygments==2.7.1
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sentry-sdk==0.19.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
win32-setctime==1.0.3
